move nitro init to separate file
